fix: changes as per review
