+++ india_compliance/patches.txt
@@ -47,4 +47,4 @@ execute:from erpnext.stock.doctype.repost_item_valuation.repost_item_valuation i
 india_compliance.patches.v14.set_item_details_from_purchase_invoice_to_bill_of_entry
 india_compliance.patches.v14.update_item_gst_details_and_gst_trearment_in_bill_of_entry
 india_compliance.patches.v14.update_default_auto_reconciliation_settings
-india_compliance.patches.v14.delete_gstr_import_log_with_error_type_RET2B1023
+india_compliance.patches.v14.delete_not_generated_gstr_import_log
